LayoutTests/fast/dom/processing-instruction-appendChild-exceptions.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
<![CDATA[
description("Test that appropriate exceptions are thrown when adding children to a ProcessingInstruction.");
var pi = document.createProcessingInstruction('target', 'data');
shouldThrow("pi.appendChild(null)");
var div = document.createElement('div');
shouldThrow("pi.appendChild(div)", "'HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree.'");
var textNode = document.createTextNode('sometext');
shouldThrow("pi.appendChild(textNode)", "'HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree.'");
]]>
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
